added the final functionalities in register page
