+++ index.html
@@ -13,7 +13,7 @@
     <link rel="stylesheet" href="output.css" />
     <link rel="stylesheet" href="style.css" />
     <script src="script.js"></script>
-    <script src="../js/auth/login.js"></script>
+    <script src="js/auth/login.js"></script>
     <title>Sign in</title>
   </head>
   <body>
@@ -94,7 +94,7 @@
                   <div class="flex items-center justify-center">
                     <button
                       type="submit"
-                      class="rounded-md bg-indigo-600 hover:border-2 border-indigo-900 px-6 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600 mb-7"
+                      class="rounded-md bg-indigo-600 w-1/2 border-indigo-900 px-6 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600 mb-7"
                     >
                       Sign in
                     </button>
@@ -146,96 +146,5 @@
         </div>
       </div>
     </section>
-
-    <!--This is for when the screen is small
-
-    <section class="hidden">
-      <div class="flex min-h-full flex-col justify-center px-6 py-12 lg:px-8">
-        <div class="sm:mx-auto sm:w-full sm:max-w-sm">
-          <img
-            class="mx-auto h-10 w-auto"
-            src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
-            alt="Your Company"
-          />
-          <h2
-            class="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900"
-          >
-            Sign in to your account
-          </h2>
-        </div>
-
-        <div class="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
-          <form id="form_login2" class="space-y-6" action="#" method="POST">
-            <div>
-              <label
-                for="email2"
-                class="block text-sm font-medium leading-6 text-gray-900"
-                >Email address</label
-              >
-              <div class="mt-2">
-                <input
-                  id="email2"
-                  name="email2"
-                  type="email"
-                  autocomplete="email2"
-                  required
-                  class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                />
-              </div>
-            </div>
-
-            <div>
-              <div class="flex items-center justify-between">
-                <label
-                  for="password2"
-                  class="block text-sm font-medium leading-6 text-gray-900"
-                  >Password</label
-                >
-                <div class="text-sm">
-                  <a
-                    href="#"
-                    class="font-semibold text-indigo-600 hover:text-indigo-500"
-                    >Forgot password?</a
-                  >
-                </div>
-              </div>
-              <div class="mt-2 flex relative items-center">
-                <input
-                  id="password2"
-                  name="password2"
-                  type="password"
-                  required
-                  class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 pl-2"
-                />
-                <div class="flex absolute-right">
-                  <i
-                    id="togglePassword2"
-                    class="fa-solid fa-eye cursor-pointer p-2"
-                  ></i>
-                </div>
-              </div>
-            </div>
-
-            <div>
-              <button
-                type="submit"
-                class="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-              >
-                Sign in
-              </button>
-            </div>
-          </form>
-
-          <p class="mt-10 text-center text-sm text-gray-500">
-            Not a member?
-            <a
-              href="register.html"
-              class="font-semibold leading-6 text-indigo-600 hover:text-indigo-500"
-              >Sign up here!</a
-            >
-          </p>
-        </div>
-      </div>
-    </section>-->
   </body>
 </html>
