+++ js/auth/register.js
@@ -1,13 +1,13 @@
-document.addEventListener("DOMContentLoaded", () => {
+import { url, successNotification, errorNotification } from "../utils/util.js";
+
+document.addEventListener("DOMContentLoaded", () => {  
   // Here lies the ngrok URL
-  const url = "http://109.test";
+  
 
   const form_register = document.getElementById("form_register");
 
   form_register.onsubmit = async (e) => {
     e.preventDefault();
-    console.log("click is okay");
-
     //Disable the button
     document.querySelector("#form_register button").disabled = true;
     document.querySelector("#form_register button").innerHTML = `<span data-te-loading-text-ref>Loading...</span>`;  
@@ -44,27 +44,7 @@ document.addEventListener("DOMContentLoaded", () => {
     }
 
     // Enable button
-    document.querySelector("#form_register button").disabled = false;document.querySelector("#form_register button").innerHTML = `Sign up`;  
+    document.querySelector("#form_register button").disabled = false;
+    document.querySelector("#form_register button").innerHTML = `Sign up`;  
   };
-
-  function successNotification(message, seconds) {
-    document.querySelector(".alert-success").classList.remove("hidden");
-    document.querySelector(".alert-success").classList.add("block");
-    document.querySelector(".alert-success").innerHTML = message;
-
-    setTimeout(function () {
-      document.querySelector(".alert-success").classList.remove("block");
-      document.querySelector(".alert-success").classList.add("hidden");
-    }, seconds * 1000);
-  }
-  function errorNotification(message, seconds) {
-    document.querySelector(".alert-unsuccessful").classList.remove("hidden");
-    document.querySelector(".alert-unsuccessful").classList.add("block");
-    document.querySelector(".alert-unsuccessful").innerHTML = message;
-
-    setTimeout(function () {
-      document.querySelector(".alert-unsuccessful").classList.remove("block");
-      document.querySelector(".alert-unsuccessful").classList.add("hidden");
-    }, seconds * 3000);
-  }
 });
